+++ src/common/utils.js
@@ -1,27 +1,10 @@
 /* @flow */
-
-// These are domain-specific
-
-import type { IStickValue } from './types';
-
-export function getRawGamepads(): Gamepad[] {
-    if (navigator && navigator.getGamepads) {
-        return Array.from(navigator.getGamepads());
-    }
-    return [];
-}
-
-export function nameIsValid(name: string) {
-    return /^[a-z0-9]+$/i.test(name);
-}
-
-export function isButtonSignificant(value: number = 0, threshold: number): boolean {
-    return Math.abs(value) > threshold;
-}
-
-export function isStickSignificant(stickValue: IStickValue, threshold: number): boolean {
-    return stickValue.findIndex(value => Math.abs(value) >= threshold) !== -1;
-}
+import type {
+    IStickState, IStickIndexes,
+    IStickValue, IStickInverts,
+    IButtonState, IButtonIndexes,
+    IGamepad
+} from './types';
 
 // These are generic Lodash function replacements
 
@@ -232,4 +215,97 @@ export function findKey(search: FindKeyCb | Object, target: Object): string | nu
     }
 
     return null;
+}
+
+
+// These are domain-specific
+
+
+export function getRawGamepads(): Gamepad[] {
+    if (navigator && navigator.getGamepads) {
+        return Array.from(navigator.getGamepads());
+    }
+    return [];
+}
+
+export function nameIsValid(name: string) {
+    return /^[a-z0-9]+$/i.test(name);
+}
+
+export function isButtonSignificant(value: number = 0, threshold: number): boolean {
+    return Math.abs(value) > threshold;
+}
+
+export function isStickSignificant(stickValue: IStickValue, threshold: number): boolean {
+    return stickValue.findIndex(value => Math.abs(value) >= threshold) !== -1;
+}
+
+export function buttonMap(
+    pad: IGamepad,
+    prevPad: IGamepad,
+    indexes: IButtonIndexes,
+    threshold: number
+): IButtonState {
+    const length = indexes.length;
+
+    let prevPressed = false;
+    let value = 0;
+    let pressed = false;
+
+    let i = 0;
+    while (i < length) {
+        if (!prevPressed) {
+            const prevValue = prevPad.buttons[indexes[i]] || 0;
+            prevPressed = isButtonSignificant(prevValue, threshold);
+        }
+
+        const currValue = pad.buttons[indexes[i]] || 0;
+        value = Math.max(value, currValue);
+        if (!pressed) {
+            pressed = isButtonSignificant(currValue, threshold);
+        }
+
+        i += 1;
+    }
+
+    return {
+        value,
+        pressed,
+        justChanged: pressed !== prevPressed
+    };
+}
+
+export function roundSticks(indexMaps: IStickIndexes, axes: number[], threshold: number): IStickValue {
+    let count = 0;
+    let counts = [];
+
+    indexMaps.forEach(indexes => {
+        const values = indexes.map(i => axes[i]);
+
+        if (isStickSignificant(values, threshold)) {
+            counts = values.map((v, i) => v + (counts[i] || 0));
+            count += 1;
+        }
+    });
+
+    return count === 0 ? indexMaps[0].map(() => 0) : counts.map(v => v / count);
+}
+
+export function stickMap(
+    pad: IGamepad,
+    prevPad: IGamepad,
+    indexMaps: IStickIndexes,
+    inverts: IStickInverts,
+    threshold: number
+): IStickState {
+    const prevPressed = isStickSignificant(roundSticks(indexMaps, prevPad.axes, threshold), threshold);
+    const value = roundSticks(indexMaps, pad.axes, threshold).map((v, i) => (!inverts[i] ? v : v * -1));
+    const pressed = isStickSignificant(value, threshold);
+
+    return {
+        value,
+        pressed,
+        justChanged: pressed !== prevPressed,
+        inverts
+    };
 }
